42 add update video and delete video

+++ src/Controller/FigureController.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\String\Slugger\AsciiSlugger;
 use App\Entity\Image;
+use App\Entity\Video;
 use App\Entity\Figure;
 use App\Entity\Connect;
 use App\Entity\Comment;
@@ -20,8 +21,9 @@ use App\Form\UpdateFigureType;
 class FigureController extends AbstractController
 {
     public function __construct(
-        private EntityManagerInterface $entityManager){
-            
+        private EntityManagerInterface $entityManager)
+    {
+
     }
 
     #[Route('/figure/{slug}', name: 'detail_figure')]
@@ -73,16 +75,20 @@ class FigureController extends AbstractController
     ): Response {
 
         //self::unknow($figure);
+        if (!is_null($this->getUser())) {
+            if ($figure->getImage()[0]) {
+                $img->delete($figure->getImage()[0]->getName(), $figure->getTitle());
+            }
 
-        if ($figure->getImage()[0]) {
-            $img->delete($figure->getImage()[0]->getName(), $figure->getTitle());
+            $this->entityManager->remove($figure);
+            $this->entityManager->flush();
+
+            $this->addFlash('success', 'Tricks supprimée avec succès');
+
+
+        } else {
+            $this->addFlash('error', ('Veuillez vous connecter'));
         }
-
-        $this->entityManager->remove($figure);
-        $this->entityManager->flush();
-
-        $this->addFlash('success', 'Tricks supprimée avec succès');
-
         return $this->redirectToRoute('app_home');
     }
 
@@ -95,36 +101,56 @@ class FigureController extends AbstractController
         UpdateFigureType $updateForm,
     ): Response {
 
+        if (!is_null($this->getUser())) {
+            $updateForm = $this->createForm(UpdateFigureType::class, $figure);
+            $updateForm->handleRequest($request);
 
-        $updateForm = $this->createForm(UpdateFigureType::class, $figure);
-        $updateForm->handleRequest($request);
 
-        //self::unknow($figure);
+            if ($updateForm->isSubmitted() && $updateForm->isValid()) {
+                $slugger = new AsciiSlugger();
+                $slug = $slugger->slug($figure->getTitle());
 
-        if ($updateForm->isSubmitted() && $updateForm->isValid()) {
-            $slugger = new AsciiSlugger();
-            $slug = $slugger->slug($figure->getTitle());
+                $images = $updateForm->get('image')->getData();
+                $videos = $updateForm->get('videos')->getData();
 
-            $images = $updateForm->get('image')->getData();
+                foreach ($images as $image) {
+                    $folder = $figure->getTitle();
 
-            foreach ($images as $image) {
-                $folder = $figure->getTitle();
+                    $fichier = $img->addImg($image, $folder, 300, 300);
+                    $picture = new Image();
+                    $picture->setName($fichier);
+                    $figure->addImage($picture);
+                }
 
-                $fichier = $img->addImg($image, $folder, 300, 300);
+                if($videos){
+                    $control = explode("/", $videos);
+                    if($control[3] != 'embed')
+                    {
+                        $addEmbed = str_replace($control[2], $control[2] . '/embed', $videos);
+                        $video = new Video();
+                        $video->setName($addEmbed);
+                        $figure->addVideo($video);
+                    }else{
+                    $video = new Video();
+                    $video->setName($videos);
+                    $figure->addVideo($video);
+                    }
+                }
 
-                $picture = new Image();
-                $picture->setName($fichier);
-                $figure->addImage($picture);
+
+                $figure->setTitle($figure->getTitle())
+                    ->setDescription($figure->getDescription())
+                    ->setDateUpdate(new \DateTime())
+                    ->setSlug($slug);
+                $this->entityManager->persist($figure);
+                $this->entityManager->flush();
+                
+
             }
+        } else {
 
-
-            $figure->setTitle($figure->getTitle())
-                ->setDescription($figure->getDescription())
-                ->setDateUpdate(new \DateTime())
-                ->setSlug($slug);
-            $this->entityManager->persist($figure);
-            $this->entityManager->flush();
-
+            $this->addFlash('error', 'Veuillez vous connecter');
+            return $this->redirectToRoute('app_home');
         }
 
         return $this->render('figure/update.html.twig', [
@@ -139,10 +165,10 @@ class FigureController extends AbstractController
         Image $image,
         Figure $figure,
     ): Response {
-      
+
         $this->entityManager->remove($image);
         $this->entityManager->flush();
-        
+
         $success = $img->deleteOne($image, $figure->getTitle());
 
         if ($success) {
@@ -151,17 +177,19 @@ class FigureController extends AbstractController
             $this->addFlash('error', 'Echec de la suppresion');
         }
 
-        return $this->redirectToRoute('update_figure', [ 'figure' => $figure->getId() ]);
+        return $this->redirectToRoute('update_figure', ['figure' => $figure->getId()]);
 
     }
+    
+    #[route(path:'deleteVideo/{figure}/{id}', name: 'delete_video')]
+    public function deleteVideo(Figure $figure, Video $video): Response
+{
+    $this->entityManager->remove($video);
+    $this->entityManager->flush();
+    
+   $this->addFlash('success','Video supprimée avec succès');
 
-    public function unknow(
-        $figure
-    ): void {
-        if (!$figure) {
-            $this->createNotFoundException(
-                'Pas de figure avec l\' ' . $figure->getId()
-            );
-        }
-    }
+   return $this->redirectToRoute('update_figure', ['figure' => $figure->getId()]);
+}
+
 }
